+++ src/collective/liveblog/browser/update.py
@@ -1,11 +1,16 @@
 # -*- coding: utf-8 -*-
 from collective.liveblog import _
 from collective.liveblog.browser.view import View
+from collective.liveblog.config import BATCH_SIZE
+from collective.liveblog.config import ORPHAN
 from collective.liveblog.interfaces import IBrowserLayer
 from collective.liveblog.interfaces import ILiveblog
 from five import grok
+from plone import api
+from plone.batching import Batch
 from zope.i18n import translate
 from zope.security import checkPermission
+
 
 grok.templatedir('templates')
 
@@ -18,6 +23,17 @@ class Update(View):
     grok.layer(IBrowserLayer)
     grok.require('collective.liveblog.AddMicroUpdate')
 
+    def update(self):
+        self.start = int(self.request.get('b_start', 0))
+        if self.start != 0:
+            msg = _(u'You must be on the first page of the batch to add micro-updates.')
+            api.portal.show_message(msg, self.request, type='info')
+
+    @property
+    def batch(self):
+        """Encapsulate sequence in batches of size."""
+        return Batch(self.updates(), BATCH_SIZE, self.start, orphan=ORPHAN)
+
     def can_edit_objects(self):
         return checkPermission('cmf.ModifyPortalContent', self.context)
 
@@ -28,3 +44,11 @@ class Update(View):
         msg = _(u'Do you really want to delete this item?')
         msg = translate(msg, 'collective.liveblog', context=self.request)
         return u"return confirm('{0}')".format(msg)
+
+    @property
+    def automatic_updates_enabled(self):
+        """Check if the Livelog must be updated automatically.
+        Automatic updates should be enabled on first page of batch.
+        """
+        enabled = super(Update, self).automatic_updates_enabled
+        return enabled and self.start == 0